+++ include/IqrfTraceHex.h
@@ -4,8 +4,8 @@
 #include <iomanip>
 
 /// formats
-#define MEM_HEX_CHAR(ptr, len)  iqrf::TracerMemHexChar((const void*)ptr, len, ' ')
-#define MEM_HEX(ptr, len)       iqrf::TracerMemHex((const void*)ptr, len, '.')
+#define MEM_HEX_CHAR(ptr, len)  iqrf::TracerMemHexChar((const void*)ptr, static_cast<long>(len), ' ')
+#define MEM_HEX(ptr, len)       iqrf::TracerMemHex((const void*)ptr, static_cast<long>(len), '.')
 
 namespace iqrf {
   class TracerMemHex
